+++ OneByte/Source/Features/Work/View/WorkView.swift
@@ -8,12 +8,27 @@
 import SwiftUI
 
 struct WorkView: View {
-    
-    @State private var navigationManager = NavigationManager()
-    
+
     var body: some View {
         VStack {
-            Text("할 일 View")
+            HStack {
+                Text("할 일")
+                    .font(.Pretendard.Bold.size22)
+                    .foregroundStyle(Color(hex: "B4A99D"))
+                
+                Spacer()
+                
+                NavigationLink {
+                    SettingView()
+                } label: {
+                    Image(systemName: "info.circle")
+                        .resizable()
+                        .frame(width: 24, height: 24)
+                        .foregroundStyle(Color(hex: "B4A99D"))
+                }
+            }
+            .padding()
+            Spacer()
         }
     }
 }
